#17 [Refactor]: MapScreen 관련 view 파일 분리
